add opentelemetry_header_mask directive
The directive enables regex-based replacement of sensitive parts of
header values in the span logs with specified mask. A regular expression can contain
named and positional capturing groups that can later be used in the mask
value. The directive may be specified multiple times for a particular
header. In this case, the replacement will be performed in a chain in
the order of specification.

Syntax: opentelemetry_header_mask header regex mask;

Example: opentelemetry_header_mask Authorization ^.*$ SECRET;